requirements.txt
annotated-types==0.7.0
anyio==4.11.0
bcrypt==4.0.1
click==8.3.0
ecdsa==0.19.1
fastapi==0.115.0
greenlet==3.2.4
h11==0.16.0
idna==3.11
passlib==1.7.4
pyasn1==0.6.1
pydantic==2.7.4
pydantic_core==2.18.4
python-dotenv==1.1.1
python-jose==3.3.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.44
starlette==0.38.6
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.30.6
